Enum martin::ResourceRecord

pub enum ResourceRecord {
    A {
        name: Name,
        class: Class,
        ttl: i32,
        addr: Ipv4Addr,
    },
    AAAA {
        name: Name,
        class: Class,
        ttl: i32,
        addr: Ipv6Addr,
    },
    CNAME {
        name: Name,
        class: Class,
        ttl: i32,
        cname: Name,
    },
    SOA {
        name: Name,
        class: Class,
        ttl: i32,
        mname: Name,
        rname: Name,
        serial: u32,
        refresh: u32,
        retry: u32,
        expire: u32,
        minimum: u32,
    },
    PTR {
        name: Name,
        class: Class,
        ttl: i32,
        ptrname: Name,
    },
    MX {
        name: Name,
        class: Class,
        ttl: i32,
        preference: u16,
        exchange: Name,
    },
    NS {
        name: Name,
        class: Class,
        ttl: i32,
        ns_name: Name,
    },
    OPT {
        payload_size: u16,
        extended_rcode: u8,
        version: u8,
        dnssec_ok: bool,
        data: Vec<u8>,
    },
    TXT {
        name: Name,
        class: Class,
        ttl: i32,
        data: Vec<String>,
    },
    Unknown {
        name: Name,
        rtype: Type,
        class: Class,
        ttl: i32,
        data: Vec<u8>,
    },
}

A resource record associates a Name within a Class with Type dependent data.

Variants
A

An IPv4 host address resource record.

Fields of A

name: Name

The Name this record applies to.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

addr: Ipv4Addr

The IPv4 host address.

AAAA

An IPv6 host address resource record.

Fields of AAAA

name: Name

The Name this record applies to.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

addr: Ipv6Addr

The IPv6 host address.

CNAME

The canonical name for an alias.

Fields of CNAME

name: Name

The Name this record applies to.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

cname: Name

The canonical name for the alias referred to in name.

SOA

The start of a zone of authority.

Fields of SOA

name: Name

The Name this record applies to.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

mname: Name

The of the name server that was the original or primary source of data for this zone.

rname: Name

A which specifies the mailbox of the person responsible for this zone.

serial: u32

The unsigned 32 bit version number of the original copy of the zone.

Zone transfers preserve this value. This value wraps and should be compared using sequence space arithmetic.

refresh: u32

A 32 bit time interval before the zone should be refreshed.

retry: u32

A 32 bit time interval that should elapse before a failed refresh should be retried.

expire: u32

A 32 bit time value that specifies the upper limit on the time interval that can elapse before the zone is no longer authoritative.

minimum: u32

The unsigned 32 bit minimum TTL field that should be exported with any RR from this zone.

PTR

Pointer to a canonical name.

Fields of PTR

name: Name

The Name this record applies to.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

ptrname: Name

The canonical name pointed to in name.

MX

Mail Exchange information.

Fields of MX

name: Name

The Name this record applies to.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

preference: u16

The preference given to this RR - lower values are preferred.

exchange: Name

A host willing to act as a mail exchange for the owner name.

NS

An authoritative name server.

Fields of NS

name: Name

The Name this record applies to.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

ns_name: Name

A host which should be authoritative for the specified class and domain.

OPT

A pseudo-record containing additional EDNS(0) information.

Fields of OPT

payload_size: u16

The requestor's UDP payload size.

extended_rcode: u8

An extended response code.

version: u8

The specification version supported.

dnssec_ok: bool

The DNSSEC OK bit.

data: Vec<u8>

Additional data in the form of attribute, value pairs.

TXT

Text string record information.

Fields of TXT

name: Name

The Name this record applies to.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

data: Vec<String>

One or more character strings.

Unknown

A yet-unknown type of resource record.

Fields of Unknown

name: Name

The Name this record applies to.

rtype: Type

The type code for this unknown data.

class: Class

The Class this record applies to.

ttl: i32

The "time to live" for this data, in seconds.

data: Vec<u8>

The data contained by the unknown record type.