Enum martin::Type

pub enum Type {
    A,
    AAAA,
    CNAME,
    PTR,
    SOA,
    OPT,
    MX,
    NS,
    TXT,
    Unknown {
        value: u16,
    },
}

A Type field indicates the structure and content of a resource record.

Variants
A

The A resource type, holding an IPv4 host address resource record.

AAAA

The A resource type, holding an IPv6 host address resource record.

CNAME

The CNAME resource type, holding the canonical name for an alias.

PTR

The PTR resource type, pointing to a canonical name. Does not trigger CNAME processing.

SOA

The SOA resource type, marks the start of a zone of authority.

OPT

The OPT pseudo-RR type, adding additional EDNS(0) information to a request / response.

MX

The MX resource type, holding mail exchange information.

NS

The NS resource type, holding an authoritative name server.

TXT

The TXT resource type, holding text strings.

Unknown

Indicates that the type is not known to this parser.

Fields of Unknown

value: u16

The value of the unknown type